feat: refactor authentication and database handling; remove old models and config
