+++ backend/internal/database/db.go
@@ -1,0 +1,54 @@
+package database
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	_ "github.com/lib/pq"
+)
+
+var DB *sql.DB
+
+func Connect() {
+	connStr := fmt.Sprintf("host=%s user=%s password=%s dbname=%s port=5432 sslmode=disable",
+		os.Getenv("POSTGRES_HOST"),
+		os.Getenv("POSTGRES_USER"),
+		os.Getenv("POSTGRES_PASSWORD"),
+		os.Getenv("POSTGRES_DB"),
+	)
+
+	var err error
+	DB, err = sql.Open("postgres", connStr)
+	if err != nil {
+		log.Fatal(err)
+	}
+	if err = DB.Ping(); err != nil {
+		log.Fatal("Could not connect to database:", err)
+	}
+
+	DB.SetMaxOpenConns(25)
+	DB.SetMaxIdleConns(5)
+	DB.SetConnMaxLifetime(5 * time.Minute)
+}
+
+func InitSchema() {
+	query := `
+	CREATE TABLE IF NOT EXISTS users (
+		id UUID PRIMARY KEY,
+		email TEXT UNIQUE NOT NULL,
+		password_hash TEXT NOT NULL
+	);
+	CREATE TABLE IF NOT EXISTS todos (
+		id UUID PRIMARY KEY,
+		item TEXT NOT NULL,
+		completed BOOLEAN DEFAULT FALSE,
+		user_id UUID REFERENCES users(id)
+	);`
+	_, err := DB.Exec(query)
+	if err != nil {
+		log.Fatal("Failed to init db:", err)
+	}
+}
