--- backend/internal/models/todo.go
@@ -1,8 +1,0 @@
-package models
-
-type Todo struct {
-	ID        string `json:"id"`
-	Item      string `json:"item"`
-	Completed bool   `json:"completed"`
-	UserID    int64  `json:"-"`
-}
