--- backend/internal/models/user.go
@@ -1,7 +1,0 @@
-package models
-
-type User struct {
-	ID           int64  `json:"id"`
-	Email        string `json:"email"`
-	PasswordHash string `json:"-"`
-}
